Made `IPrint3dFan.Voltage` `double`

+++ src/Print3dServer.Core/Interfaces/IPrint3dFan.cs
@@ -4,7 +4,7 @@
     {
         #region Properties
         public bool On { get; set; }
-        public long? Voltage { get; set; }
+        public double? Voltage { get; set; }
         public int? Speed { get; }
         public int? Percent { get; }
         #endregion
